It appears to run just fine. See #46.
